Add Royal Chapel

+++ src/mapper.py
@@ -5,6 +5,7 @@ import datetime
 import hashlib
 import json
 import os
+import pathlib
 import random
 import yaml
 
@@ -82,10 +83,10 @@ class Room:
     
     def get_cells(self, offset_top: int=0, offset_left: int=0) -> set[tuple[int, int]]:
         result = set()
-        for row in range(self.top, self.top + self.rows):
-            for col in range(self.left, self.left + self.columns):
+        for row in range(self.rows):
+            for col in range(self.columns):
                 if (row, col) not in self.empty_cells:
-                    result.add((row + offset_top, col + offset_left))
+                    result.add((self.top + row + offset_top, self.left + col + offset_left))
         return result
 
 class RoomSet:
@@ -433,8 +434,8 @@ stages = {
     'Long Library': [
         {
             'Long Library, Exit to Outer Wall': (32 + 0, 32 + 0),
-            'Long Library, Loading Room A': (32 + 0, 32 + 3),
-            'Long Library, Fake Room With Teleporter A': (32 + 0, 32 + 4),
+            'Long Library, Loading Room A': (32 + 2, 32 + 3),
+            'Long Library, Fake Room With Teleporter A': (32 + 2, 32 + 4),
         },
         {
             'Long Library, Spellbook Area': (0, 0),
@@ -526,6 +527,45 @@ stages = {
         { 'Castle Keep, Save Room A': (0, 0) },
         { 'Castle Keep, Tyrfing Room': (0, 0) },
     ],
+    'Royal Chapel': [
+        {
+            # NOTE(sestren): For now, these hallways and towers must be combined until the special behavior that controls transitions between them is better understood
+            'Royal Chapel, Spike Hallway': (32 + 4, 32 + 0),
+            'Royal Chapel, Left Tower': (32 + 1, 32 + 3),
+            'Royal Chapel, Walkway Between Towers': (32 + 3, 32 + 5),
+            'Royal Chapel, Middle Tower': (32 + 0, 32 + 8),
+            'Royal Chapel, Walkway Left of Hippogryph': (32 + 2, 32 + 10),
+        },
+        {
+            'Royal Chapel, Walkway Right of Hippogryph': (0 + 3, 0 + 0),
+            'Royal Chapel, Right Tower': (0 + 0, 0 + 1),
+            'Royal Chapel, Loading Room A': (0 + 2, 0 + 4),
+            'Royal Chapel, Fake Room With Teleporter A': (0 + 2, 0 + 5),
+        },
+        {
+            'Royal Chapel, Pushing Statue Shortcut': (0 + 0, 0 + 0),
+            'Royal Chapel, Loading Room D': (0 + 0, 0 + 1),
+            'Royal Chapel, Fake Room With Teleporter B': (0 + 0, 0 + 2),
+        },
+        {
+            'Royal Chapel, Nave': (0 + 0, 0 + 0),
+            'Royal Chapel, Loading Room C': (0 + 1, 0 + 2),
+            'Royal Chapel, Fake Room With Teleporter C': (0 + 1, 0 + 3),
+        },
+        {
+            'Royal Chapel, Statue Ledge': (0 + 0, 0 + 0),
+            'Royal Chapel, Loading Room B': (0 + 0, 0 + 1),
+            'Royal Chapel, Fake Room With Teleporter D': (0 + 0, 0 + 2),
+        },
+        { 'Royal Chapel, Chapel Staircase': (0, 0) },
+        { 'Royal Chapel, Confessional Booth': (0, 0) },
+        { 'Royal Chapel, Empty Room': (0, 0) },
+        { 'Royal Chapel, Goggles Room': (0, 0) },
+        { 'Royal Chapel, Hippogryph Room': (0, 0) },
+        { 'Royal Chapel, Save Room A': (0, 0) },
+        { 'Royal Chapel, Save Room B': (0, 0) },
+        { 'Royal Chapel, Silver Ring Room': (0, 0) },
+    ],
 }
 
 def get_roomset(rng, rooms: dict, stage_data: dict) -> RoomSet:
@@ -545,7 +585,7 @@ def get_roomset(rng, rooms: dict, stage_data: dict) -> RoomSet:
     while len(pool) > 0:
         possible_target_nodes = result.get_open_nodes()
         if len(possible_target_nodes) < 1:
-            # ERROR: No open nodes left
+            # print('ERROR: No open nodes left')
             break
         target_node = rng.choice(possible_target_nodes)
         open_nodes = []
@@ -553,7 +593,7 @@ def get_roomset(rng, rooms: dict, stage_data: dict) -> RoomSet:
             for open_node in roomset.get_open_nodes(matching_node=target_node):
                 open_nodes.append(open_node)
         if len(open_nodes) < 1:
-            # ERROR: No matching source nodes for the chosen target node
+            # print('ERROR: No matching source nodes for the chosen target node')
             break
         # Go through possible source nodes in random order until a valid source node is found
         open_nodes.sort()
@@ -567,7 +607,7 @@ def get_roomset(rng, rooms: dict, stage_data: dict) -> RoomSet:
                 roomset = pool.pop(roomset_key, None)
                 break
         else:
-            # ERROR: All matching source nodes for the target node result in invalid room placement
+            # print('ERROR: All matching source nodes for the target node result in invalid room placement')
             break
         steps += 1
     return result
@@ -589,6 +629,7 @@ class MapperData:
             'clock-tower',
             'warp-rooms',
             'castle-keep',
+            'royal-chapel',
         ):
             folder_path = os.path.join('data', 'rooms', stage_folder)
             for file_name in os.listdir(folder_path):
@@ -613,7 +654,24 @@ class MapperData:
 class LogicCore:
     def __init__(self, mapper_data, changes):
         print('Build logic core')
-        self.commands = {}
+        self.commands = {
+            'Global': {
+                'Use Library Card': {
+                    'Outcomes': {
+                        'Location': 'Long Library, Outside Shop',
+                        'Section': 'Main',
+                        'Item - Library Card': -1,
+                    },
+                    'Requirements': {
+                        'Default': {
+                            'Item - Library Card': {
+                                'Minimum': 1,
+                            },
+                        },
+                    },
+                },
+            },
+        }
         for stage_name in (
             'Castle Entrance',
             'Castle Entrance Revisited',
@@ -626,6 +684,7 @@ class LogicCore:
             'Clock Tower',
             'Warp Rooms',
             'Castle Keep',
+            'Royal Chapel',
         ):
             # print('', stage_name)
             nodes = {}
@@ -834,7 +893,7 @@ class Mapper:
             legend.append((code, room_name))
             for cell_row in range(max(0, room.top), min(64, room.top + room.rows)):
                 for cell_col in range(max(0, room.left), min(64, room.left + room.columns)):
-                    if (cell_row - stage_top, cell_col - stage_left) in room.empty_cells:
+                    if (cell_row - room.top, cell_col - room.left) in room.empty_cells:
                         continue
                     top = cell_row - stage_top
                     left = cell_col - stage_left
@@ -869,8 +928,16 @@ if __name__ == '__main__':
     Usage
     python mapper.py
     '''
+    for stage_name in stages:
+        pathlib.Path(
+            os.path.join('build', 'mapper', stage_name)
+        ).mkdir(parents=True, exist_ok=True)
     GENERATION_VERSION = '0.0.4'
     mapper_core = MapperData().get_core()
+    with (
+        open(os.path.join('build', 'mapper', 'mapper-core.json'), 'w') as mapper_core_json,
+    ):
+        json.dump(mapper_core, mapper_core_json, indent='    ', sort_keys=True, default=str)
     parser = argparse.ArgumentParser()
     parser.add_argument('stage_name', help='Input a valid stage name', type=str)
     parser.add_argument('stage_count', help='Input the number of stage instances to generate', type=int)
@@ -885,12 +952,13 @@ if __name__ == '__main__':
             if stage_map.validate():
                 break
         changes = stage_map.stage.get_changes()
+        hash_of_rooms = hashlib.sha256(json.dumps(changes['Rooms'], sort_keys=True).encode()).hexdigest()
         mapper_data = {
             'Attempts': stage_map.attempts,
             'Generation Start Date': stage_map.start_time.isoformat(),
             'Generation End Date': stage_map.end_time.isoformat(),
             'Generation Version': GENERATION_VERSION,
-            'Hash of Rooms': hashlib.sha256(json.dumps(changes['Rooms'], sort_keys=True).encode()).hexdigest(),
+            'Hash of Rooms': hash_of_rooms,
             'Seed': stage_map.current_seed,
             'Stage': args.stage_name,
         }
@@ -900,8 +968,12 @@ if __name__ == '__main__':
         # spoiler = stage_map.get_spoiler(args.stage_name)
         # for line in spoiler:
         #     print(line)
-        with (
-            open(os.path.join('build', 'mapper', args.stage_name, str(stage_map.current_seed) + '.json'), 'w') as mapper_data_json,
-        ):
-            json.dump(mapper_data, mapper_data_json, indent='    ', sort_keys=True, default=str)
+        filepath = os.path.join('build', 'mapper', args.stage_name, hash_of_rooms + '.json')
+        if not os.path.exists(filepath):
+            with (
+                open(os.path.join('build', 'mapper', args.stage_name, hash_of_rooms + '.json'), 'w') as mapper_data_json,
+            ):
+                json.dump(mapper_data, mapper_data_json, indent='    ', sort_keys=True, default=str)
+        else:
+            print('Stage with that hash already exists:', (args.stage_name, hash_of_rooms))
         seed = stage_map.rng.randint(0, 2 ** 64)
